requirements.txt
streamlit==1.29.0
crewai==0.28.8
crewai-tools==0.1.6
langchain-community==0.0.29
langchain-openai==0.0.5
openai==1.12.0
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0
